the night in brahim room i did a lot of linking but i hope its not fucked

+++ lib/features/client/orderhisyory/data/ordermodel.dart
@@ -1,10 +1,10 @@
-import 'package:project_2cp/features/client/orderlist/data/ordermodel.dart';
+import 'package:project_2cp/features/client/orderlist/data/Item_mdeol.dart';
 
 class OrderHistory {
   final String date;
   final String time;
   final int state;
-  final int price;
+  final double price;
   OrderHistory({
     required this.date,
     required this.price,
@@ -14,11 +14,7 @@ class OrderHistory {
   factory OrderHistory.fromOrder(Item order) {
     return OrderHistory(
       time:"HH-MM-SS" ,
-      
-      
       price: order.price,
-      
-      
       date: "DD-MM-YYYY",
       state: 0, // Default status
     );
