+++ lib/features/client/orderlist/providers/totalprovider.dart
@@ -2,7 +2,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:project_2cp/features/client/orderlist/providers/listprovider.dart';
 import 'package:project_2cp/features/client/orderlist/providers/orderprovider.dart';
 
-final totalProvider = Provider<int>((ref) {
+final totalProvider = Provider<double>((ref) {
   final orders = ref.watch(orderListProvider);
   final quantities = ref.watch(amountProvider);
